api/requirements.txt
azure-functions
azure-storage-blob
pydantic>=2.0
openai
httpx
rapidfuzz
python-dotenv
